Backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
python-jose[cryptography]
passlib[bcrypt]